requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
colorama==0.4.1
cryptography==2.7
emoji==0.5.3
future==0.17.1
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
PySocks==1.7.0
python-telegram-bot==12.0.0
six==1.12.0
tornado==6.0.3
typed-ast==1.4.0
wrapt==1.11.2
clarifai==2.6.2
